requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
twilio==8.10.0
python-dotenv==1.0.1
schedule==1.2.1
pytest==8.1.0